resturctured the project, archieved unused files by server and commenting out the code in those files
