+++ package.json
@@ -2,11 +2,10 @@
   "name": "phiinternal",
   "version": "1.0.0",
   "description": "Internal project for WLU Phi",
-  "main": "server.js",
   "scripts": {
     "client-install": "npm install --prefix client",
     "start": "node server.js",
-    "server": "nodemon server.js",
+    "server": "npm start --prefix server",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "lint": "eslint models/** routes/** validation/** server.js"
@@ -14,20 +13,6 @@
   "author": "WLU Phi",
   "license": "MIT",
   "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
-    "concurrently": "^5.3.0",
-    "express": "^4.17.1",
-    "is-empty": "^1.2.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.17",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "validator": "^13.1.17"
-  },
-  "devDependencies": {
-    "eslint": "^7.18.0",
-    "eslint-plugin-react": "^7.22.0",
-    "nodemon": "^2.0.6"
+    "concurrently": "^5.3.0"
   }
 }
