+++ server/server.js
@@ -1,8 +1,6 @@
 const express = require("express");
 const mongoose = require("mongoose");
 const bodyParser = require("body-parser");
-const passport = require("passport");
-const users = require("./routes/api/users");
 const jobs = require("./routes/api/jobs");
 
 const app = express();
@@ -16,18 +14,10 @@ app.use(bodyParser.json());
 // DB Config
 const db = require("./config/keys").mongoURI;
 // Connect to MongoDB
-mongoose
-  .connect(db, { useNewUrlParser: true })
-  .then(() => console.log("MongoDB successfully connected"))
-  .catch((err) => console.log(err));
-
-// Passport middleware
-app.use(passport.initialize());
-// Passport config
-require("./config/passport")(passport);
-// Routes
-app.use("/api/users", users);
-app.use("api/jobs", jobs);
+// mongoose
+//   .connect(db, { useNewUrlParser: true })
+//   .then(() => console.log("MongoDB successfully connected"))
+//   .catch((err) => console.log(err));
 
 const port = process.env.PORT || 5000; // process.env.port is Heroku's port if you choose to deploy the app there
 app.listen(port, () => console.log(`Server up and running on port ${port} !`));
